# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Shopify API",
        "description": "RESTful API for managing Shopify products and variants with GraphQL integration",
        "contact": {
            "name": "API Support",
            "email": "dev@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Development API Server"
        }
    ],
    "paths": {
...
        "/api/v1/search/variants": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Advanced variant search",
                "description": "Search variants with multiple filters",
                "operationId": "6bb549d2deeceb6a5816719fae03760b",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search query for SKU, options",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "Filter by product ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "description": "Minimum price filter",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "decimal"
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "description": "Maximum price filter",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "decimal"
                        }
                    },
                    {
                        "name": "has_inventory",
                        "in": "query",
                        "description": "Filter by inventory availability",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Variant search results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Variant"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "count": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "current_page": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/variants": {
            "get": {
                "tags": [
                    "Variants"
                ],
                "summary": "Get paginated list of variants",
                "description": "Retrieve variants with optional filtering and search",
                "operationId": "093380c8a2fbba8accfd507c05a8d6fa",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100
                        }
                    },
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "Filter by product ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search variants by SKU or title",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "description": "Minimum price filter",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "decimal"
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "description": "Maximum price filter",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "decimal"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Variant"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "last_page": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
        "/api/v1/products/{product}/variants": {
            "get": {
                "tags": [
                    "Variants"
                ],
                "summary": "Get variants for a product",
                "description": "Retrieve all variants for a specific product",
                "operationId": "6cbbd9ecb3d6b653318df6e9c0617965",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Variant"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Variants"
                ],
                "summary": "Create a new variant",
                "description": "Create a new variant for a product and sync to Shopify",
                "operationId": "489c1d777f477e636cdafd6a7e4d23ba",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "option1": {
                                        "description": "First variant option",
                                        "type": "string",
                                        "example": "Medium",
                                        "nullable": true
                                    },
                                    "option2": {
                                        "description": "Second variant option",
                                        "type": "string",
                                        "example": "Red",
                                        "nullable": true
                                    },
                                    "price": {
                                        "description": "Variant price (required)",
                                        "type": "number",
                                        "format": "decimal",
                                        "minimum": 0,
                                        "example": 29.99
                                    },
                                    "sku": {
                                        "description": "Stock keeping unit",
                                        "type": "string",
                                        "example": "TSHIRT-M-RED",
                                        "nullable": true
                                    },
                                    "inventory_quantity": {
                                        "description": "Available inventory quantity",
                                        "type": "integer",
                                        "minimum": 0,
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Variant created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Variant"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal Server Error - Shopify sync failed"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Variant": {
                "title": "Product Variant",
                "description": "Product variant with pricing and inventory",
                "properties": {
                    "id": {
                        "description": "Local variant ID",
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "description": "Parent product ID",
                        "type": "integer",
                        "example": 1
                    },
                    "shopify_variant_id": {
                        "description": "Shopify variant ID",
                        "type": "integer",
                        "example": 987654322,
                        "nullable": true
                    },
                    "option1": {
                        "description": "First variant option",
                        "type": "string",
                        "example": "Medium",
                        "nullable": true
                    },
                    "option2": {
                        "description": "Second variant option",
                        "type": "string",
                        "example": "Red",
                        "nullable": true
                    },
                    "price": {
                        "description": "Variant price",
                        "type": "number",
                        "format": "decimal",
                        "example": 29.99
                    },
                    "sku": {
                        "description": "Stock keeping unit",
                        "type": "string",
                        "example": "TSHIRT-M-RED",
                        "nullable": true
                    },
                    "inventory_quantity": {
                        "description": "Available inventory quantity",
                        "type": "integer",
                        "example": 50
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-20T15:45:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum Bearer Token authentication. Use the token returned from /api/v1/auth/login endpoint in the format: Bearer {your_token}",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}